Fix unused orders range bug and add unit test
